added resolvers
