+++ graphql/resolvers/post.js
@@ -1,0 +1,14 @@
+const Post = require('../../models/Post')
+
+module.exports = {
+    Query: {
+        async getPosts() {
+            try {
+                const posts = await Post.find();
+                return posts
+            } catch (error) {
+                throw new Error(error)
+            }
+        }
+    }
+}
